Add NavigationEvents in root layout
