+++ src/app/layout.tsx
@@ -1,6 +1,9 @@
 import './globals.css';
 import * as C from 'components';
-import GoogleAnalytics from './googleAnalytics';
+import { GA_TRACKING_ID } from 'libs';
+import Script from 'next/script';
+import { Suspense } from 'react';
+
 export const metadata = {
   title: 'EveryGSM',
   description: 'Generated by Themoment',
@@ -13,10 +16,32 @@ export default function RootLayout({
 }) {
   return (
     <html lang="ko">
+      <head>
+        {/* Global Site Tag (gtag.js) - Google Analytics */}
+        <Script
+          async
+          src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`}
+        />
+        <Script
+          id="gtag-init"
+          strategy="afterInteractive"
+          dangerouslySetInnerHTML={{
+            __html: `
+              window.dataLayer = window.dataLayer || [];
+              function gtag(){dataLayer.push(arguments);}
+              gtag('js', new Date());
+              gtag('config', '${GA_TRACKING_ID}', {
+                page_path: window.location.pathname,
+              });
+            `,
+          }}
+        />
+      </head>
       <body>
         <C.Nav />
         {children}
         <C.Footer />
+        <Suspense fallback={null}></Suspense>
       </body>
     </html>
   );
